pyproject.toml
[project]
name = "trade-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.1.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg[binary]>=3.3.3",
]
